chore: clean repo and clarify entry points
