--- .github/workflows/docs.yml
@@ -1,18 +1,0 @@
-name: Build Documentation
-on:
-  push:
-    branches: [ main ]
-jobs:
-  build-docs:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
-        with:
-          python-version: 3.x
-      - run: pip install mkdocs mkdocs-material
-      - run: mkdocs build
-      - uses: peaceiris/actions-gh-pages@v3
-        with:
-          publish_dir: ./site
-
