+++ README.md
@@ -1,85 +1,21 @@
 # Signature Block Recovery
 
-This repository will contain tooling for extracting or reconstructing signature blocks from various data sources. The project is currently in its planning phase.
+Recover email signature blocks from PST files.
 
-## Features
-
-- Documentation outlining the project goals and future development steps.
-- Guidelines for Codex automation via `AGENTS.md`.
-
-## Quick-Start (GUI or CLI)
-
-### GUI Installation
+## Quick Start
 
 1. **Download the Installer**  
-   Visit our [GitHub Releases page](https://github.com/epic-serendipity/Signature-Block-Recovery/releases) and grab the latest:  
-   - `SignatureRecoverySetup.msi` for Windows  
-   - `SignatureRecovery.dmg` for macOS
+   Grab the latest release from the [Releases](https://github.com/epic-serendipity/Signature-Block-Recovery/releases) page.
 
-2. **Install**  
-   - **Windows:** Double-click the `.msi`—no admin rights needed.  
-   - **macOS:** Open the `.dmg` and drag “Signature Recovery” into Applications.
+2. **Launch the GUI**  
+   Run the installer and start the app with a single click or from a terminal:
+   ```bash
+   recover-gui
+   ```
 
-3. **Launch**  
-   - Click the desktop icon (Windows) or open from Launchpad (macOS).
-
-### CLI Installation & Usage
-
-If you prefer the command line or need scripting:
-
-```bash
-pip install signature-recovery
-```
-
-* **Extract** signatures:
-
-  ```bash
-  recover-signatures extract --input my.pst --index sigs.db
-  ```
-* **Query** the index:
-
-  ```bash
-  recover-signatures query --index sigs.db --filter "john.doe"
-  ```
-* **Export** results:
-
-  ```bash
-  recover-signatures export --index sigs.db --format csv --output signatures.csv
-  ```
-* **GUI** from CLI:
-
-  ```bash
-  recover-signatures gui --index sigs.db
-  ```
-* **Help** and flags:
-
-  ```bash
-  recover-signatures --help
-  ```
-
----
-
-## Core Architecture
-
-```mermaid
-graph TD
-    PSTParser --> SignatureExtractor
-    SignatureExtractor --> SignatureParser
-    SignatureParser --> SignatureDeduplicator
-```
-
-### Extraction Parameters
-
-Configuration is loaded from a YAML file. Copy `config.example.yaml` and adjust
-patterns or fallback line counts as needed:
-
-```yaml
-extraction:
-  max_fallback_lines: 5
-  signoff_patterns:
-    - "--\s*$"
-    - "regards"
-parser:
-  phone_patterns:
-    - "\(\d{3}\)\s*\d{3}-\d{4}"
-```
+3. **Command Line**  
+   Advanced users can work directly with the CLI:
+   ```bash
+   pip install signature-recovery
+   recover-signatures extract --input my.pst --index sigs.db
+   ```
