--- config.example.yaml
@@ -1,12 +1,0 @@
-extraction:
-  max_fallback_lines: 5
-  signoff_patterns:
-    - "--\s*$"
-    - "regards"
-    - "thanks"
-parser:
-  phone_patterns:
-    - "\(\d{3}\)\s*\d{3}-\d{4}"
-    - "\+\d{1,2}\s\d{3}\s\d{3}\s\d{4}"
-logging:
-  file: recovery.log
