--- mkdocs.yml
@@ -1,13 +1,0 @@
-site_name: Signature Block Recovery
-nav:
-  - Home: index.md
-  - Quick Start: quickstart.md
-  - CLI Usage: cli_usage.md
-  - GUI Usage: gui_usage.md
-  - API Reference: api_reference.md
-  - Troubleshooting: troubleshooting.md
-theme:
-  name: material
-markdown_extensions:
-  - admonition
-  - codehilite
