+++ requirements.txt
@@ -1,0 +1,1 @@
+pypff>=0.6.0
