--- signature_recovery.egg-info/SOURCES.txt
@@ -1,45 +1,0 @@
-README.md
-pyproject.toml
-setup.py
-template.py
-signature_recovery/__init__.py
-signature_recovery/api.py
-signature_recovery/exporter.py
-signature_recovery.egg-info/PKG-INFO
-signature_recovery.egg-info/SOURCES.txt
-signature_recovery.egg-info/dependency_links.txt
-signature_recovery.egg-info/entry_points.txt
-signature_recovery.egg-info/requires.txt
-signature_recovery.egg-info/top_level.txt
-signature_recovery/cli/__init__.py
-signature_recovery/cli/main.py
-signature_recovery/core/__init__.py
-signature_recovery/core/config.py
-signature_recovery/core/deduplicator.py
-signature_recovery/core/extractor.py
-signature_recovery/core/logging.py
-signature_recovery/core/metrics.py
-signature_recovery/core/models.py
-signature_recovery/core/parser.py
-signature_recovery/core/pst_parser.py
-signature_recovery/gui/__init__.py
-signature_recovery/gui/app.py
-signature_recovery/index/__init__.py
-signature_recovery/index/indexer.py
-signature_recovery/index/search_index.py
-tests/__init__.py
-tests/conftest.py
-tests/pypff.py
-tests/test_api.py
-tests/test_cli.py
-tests/test_confidence.py
-tests/test_config.py
-tests/test_deduplicator.py
-tests/test_error_logging.py
-tests/test_exporter.py
-tests/test_extractor.py
-tests/test_gui.py
-tests/test_metrics.py
-tests/test_parser.py
-tests/test_pipeline.py
-tests/test_pst_parser.py
