--- signature_recovery.egg-info/entry_points.txt
@@ -1,2 +1,0 @@
-[console_scripts]
-recover-signatures = signature_recovery.cli.main:main
