--- signature_recovery.egg-info/requires.txt
@@ -1,6 +1,0 @@
-flask>=2.0.0
-openpyxl>=3.0.0
-pandas>=1.0.0
-pyyaml>=5.4
-pyvirtualdisplay>=3.0
-argcomplete>=3.0
