--- signature_recovery.egg-info/top_level.txt
@@ -1,3 +1,0 @@
-signature_recovery
-template
-tests
